requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python
deepface
tf-keras
numpy
pydantic
transformers
torch
ffmpeg-python
librosa
groq
python-dotenv
